+++ src/js/product.js
@@ -1,22 +1,5 @@
-// import { setLocalStorage, getParam } from "./utils.mjs";
-// import { findProductById } from "./productData.mjs";
 import { getParam } from "./utils.mjs";
 import productDetails from "./productDetails.mjs";
 
 const productId = getParam("product");
 productDetails(productId);
-
-// function addProductToCart(product) {
-//   setLocalStorage("so-cart", product);
-// }
-
-// add to cart button event handler
-// async function addToCartHandler(e) {
-//   const product = await findProductById(e.target.dataset.id);
-//   addProductToCart(product);
-// }
-
-// add listener to Add to Cart button
-// document
-//   .getElementById("addToCart")
-//   .addEventListener("click", addToCartHandler);
